pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelgen"
version = "0.0.7"
description = "Pixel Generator"
readme = "README.md"
authors = [
    {name = "rifqanzalbina", email = "zalbinarifqan19@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Natural Language :: Indonesian",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.6"
dependencies = [
    "Pillow>=5.2.0",
    "numpy>=1.0.0"
]
[project.urls]
Homepage = "https://github.com/rifqanzalbina/product-generator"
"Bug Tracker" = "https://github.com/rifqanzalbina/product-generator/issues"

[tool.setuptools]
license-files = []
packages = ["pixelgen"]
include-package-data = true

